New publish script
This takes the same approach that I use on various gh-pages branches and
has worked well, achiving what I think is the same net result as the one
in the last script.

It creates a new git tree at the public/ folder with no history and
force pushes that onto the master branch on the remote.  It does this
without any subtree magic and instead creates a brand new repo within the
public directory, starts an orphan branch there, adds the public content,
copies over the remote url and then force pushes.

Advantages of this approach include:

* no resetting to HEAD~ which was not well defined for me after pulling
  a merged pr (a merge commit has two parents)
* no editing of version controlled files
* no odd subtree commands